[feat] 타입 수정

+++ src/main/java/com/example/aliexpress/dto/response/Product.java
@@ -9,6 +9,6 @@ public record Product(
         int priceDiscount,
         boolean isCoupon,
         String categoryName,
-        double reviewCount,
+        long reviewCount,
         double rating) {
 }
